+++ packages/drive/index.ts
@@ -1,4 +1,2 @@
 export * from './src/typings'
-export { PathPrefixer } from './src/PathPrefixer'
-export * from './src/Router'
-export * from './src/Driver'
+export * from './src/Drivers'
