+++ packages/drive/src/Drivers/LocalDriver/UrlBuilder/index.ts
@@ -1,14 +1,18 @@
 import qs from 'qs'
 import encodeurl from 'encodeurl'
 
-import { EncryptionContract, UrlBuilderContract } from '../typings'
+import { EncryptionContract, UrlBuilderContract } from '../../typings'
 
 export class UrlBuilder implements UrlBuilderContract {
   private routeParams: any[] | Record<string, any>
   private queryString: Record<string, any> = {}
   private baseUrl: string
 
-  constructor(private encryption: EncryptionContract) { }
+  constructor(
+    private encryption: EncryptionContract,
+    private domain: string | undefined
+  ) { }
+
   private suffixQueryString(url: string): string {
     if (this.queryString) {
       const encoded = qs.stringify(this.queryString)
@@ -43,6 +47,8 @@ export class UrlBuilder implements UrlBuilderContract {
     identifier: string,
     options?: { expiresIn?: string | number; purpose?: string }
   ) {
+    // TODO: Think how we define the `domain`
+    // we want to generate relative or absolute URLs
     let url: string
 
     const signature = this.encryption.verifier.sign(
